+++ com.furyplusplus.addon/Tests/Editor/CompressorPatchTests.cs
@@ -65,6 +65,17 @@ namespace FuryPlusPlus.Tests.Editor {
         }
 
         [Test]
+        public void Sub8PatchTargetsPassedActionController() {
+            RequireInstalled();
+
+            var parameters = CompressorCompat.LayerBuildLayer.GetParameters();
+            Assert.AreEqual(2, parameters.Length);
+            Assert.AreEqual(CompressorCompat.DecisionType, parameters[0].ParameterType);
+            Assert.AreEqual(CompressorCompat.ControllerManagerType, parameters[1].ParameterType);
+            Assert.AreEqual("controller", parameters[1].Name);
+        }
+
+        [Test]
         public void LanePackingInvariants_Randomized() {
             RequireInstalled();
             var rng = new Random(20260713);